+++ library/location/admin.py
@@ -1,3 +1,4 @@
 from django.contrib import admin
-
+from .models import library_location
 # Register your models here.
+admin.site.register(library_location)