+++ changelog.md
@@ -2,17 +2,51 @@
 <b>Full Changelog</b>
 
 ---
-#### v 0.0.3 -- 2024/01/29
+#### v 0.0.3 -- 2024/01/30
+###### New:
+* Make django app (location)
+* Add model (location/library_location)
+* Add model (location/library_location) to admin
+* Add location\form.py class "form_create_loc", "form_edit_loc"
 
+* Add model (managementUser/admin_library)
+* Add model (managementUser/admin_library) to admin
+* Add managementUser\form.py class "AdminLibraryForm"
 
+* API (views)
+    * location/location_create
+    * location/location_read
+    * location/location_update
+    * location/location_delete
+* urls
+    * location/create/
+    * location/
+    * location/<int:id>/
+    * location/<int:id>/delete/
+* Templates 
+    * location/location_create
+    * location/location_read
+    * location/location_update
 
+###### Change:
+* API (views):
+    * managementUser/user_update
+        * Delete admin_library if role change from 'admin' to another
+        * Add form2 if dropdown seleceted role is 'admin'
+* Templates 
+    * register\base.html
+        * Add lokasi sidebar Location
+    * managementUser/user_update
+        * Add form2
+        * Add async form2 
+            * When dropdown role is 'admin', form2 will appear
 
-
-
-
-
-
-
+* Lainnya :
+    * settings.py 
+        * Add INSTALLED_APPS ['location.apps.LocationConfig', ]
+    * library\urls.py
+        * Add path('location/', include("location.urls")),
+    
 
 ---
 #### v 0.0.2 -- 2024/01/29
@@ -49,6 +83,7 @@
 * Make initial file {changelog, pip_req}
 * Make django project
 * Make django app (register)
+* Add model (register/user)
 * Add register\form.py
 
 * API (views)